requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
authlib==1.2.1
pandas==2.1.3
numpy==1.26.2
faker==20.1.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
starlette==0.27.0